+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.6.4-wip.
+
+- allow omitting space between `//` and `coverage` in coverage ignore comments
+- allow text after coverage ignore comments
+- throw FormatException when encountering unbalanced ignore comments instead of silently erroring
+
 ## 1.6.3
 
 - Require Dart 2.18